requirements.txt
Flask==3.1.0
#Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
psycopg2-binary==2.9.10
bcrypt==4.2.1
PyJWT==2.10.1
imagekitio==4.1.0
pydantic==2.10.6
